Simple main script
